refactor: Move `fontFiles` down as this is common
