+++ dotcom-rendering/src/web/server/articleTemplate.ts
@@ -2,6 +2,7 @@ import { brandBackground, resets } from '@guardian/source-foundations';
 import he from 'he';
 import { ASSET_ORIGIN } from '../../lib/assets';
 import { getFontsCss } from '../../lib/fonts-css';
+import { getHttp3Url } from '../lib/getHttp3Url';
 
 export const articleTemplate = ({
 	title = 'The Guardian',
@@ -13,13 +14,13 @@ export const articleTemplate = ({
 	html,
 	windowGuardian,
 	gaPath,
-	fontFiles = [],
 	ampLink,
 	openGraphData,
 	twitterData,
 	keywords,
 	initTwitter = '',
 	recipeMarkup,
+	offerHttp3,
 }: {
 	title?: string;
 	description: string;
@@ -28,7 +29,6 @@ export const articleTemplate = ({
 	lowPriorityScriptTags: string[];
 	css: string;
 	html: string;
-	fontFiles?: string[];
 	windowGuardian: string;
 	gaPath: { modern: string; legacy: string };
 	ampLink?: string;
@@ -37,11 +37,30 @@ export const articleTemplate = ({
 	keywords: string;
 	initTwitter?: string;
 	recipeMarkup?: string;
+	offerHttp3: boolean;
 }): string => {
 	const favicon =
 		process.env.NODE_ENV === 'production'
 			? 'favicon-32x32.ico'
 			: 'favicon-32x32-dev-yellow.ico';
+
+	/**
+	 * Preload the following woff2 font files
+	 * TODO: Identify critical fonts to preload
+	 */
+	const fontFiles = [
+		// 'https://assets.guim.co.uk/static/frontend/fonts/guardian-headline/noalts-not-hinted/GHGuardianHeadline-Light.woff2',
+		// 'https://assets.guim.co.uk/static/frontend/fonts/guardian-headline/noalts-not-hinted/GHGuardianHeadline-LightItalic.woff2',
+		'https://assets.guim.co.uk/static/frontend/fonts/guardian-headline/noalts-not-hinted/GHGuardianHeadline-Medium.woff2',
+		'https://assets.guim.co.uk/static/frontend/fonts/guardian-headline/noalts-not-hinted/GHGuardianHeadline-MediumItalic.woff2',
+		'https://assets.guim.co.uk/static/frontend/fonts/guardian-headline/noalts-not-hinted/GHGuardianHeadline-Bold.woff2',
+		'https://assets.guim.co.uk/static/frontend/fonts/guardian-textegyptian/noalts-not-hinted/GuardianTextEgyptian-Regular.woff2',
+		// 'https://assets.guim.co.uk/static/frontend/fonts/guardian-textegyptian/noalts-not-hinted/GuardianTextEgyptian-RegularItalic.woff2',
+		'https://assets.guim.co.uk/static/frontend/fonts/guardian-textegyptian/noalts-not-hinted/GuardianTextEgyptian-Bold.woff2',
+		'https://assets.guim.co.uk/static/frontend/fonts/guardian-textsans/noalts-not-hinted/GuardianTextSans-Regular.woff2',
+		// 'http://assets.guim.co.uk/static/frontend/fonts/guardian-textsans/noalts-not-hinted/GuardianTextSans-RegularItalic.woff2',
+		'https://assets.guim.co.uk/static/frontend/fonts/guardian-textsans/noalts-not-hinted/GuardianTextSans-Bold.woff2',
+	].map((font) => (offerHttp3 ? getHttp3Url(font) : font));
 
 	const fontPreloadTags = fontFiles.map(
 		(fontFile) =>
